feat: add `withFeatureFactory()`
The `withFeatureFactory()` function allows
passing properties, methods, or signals from
a SignalStore to a feature. It is an advanced
feature, primarily targeted for library
authors for SignalStore features.

Its usage is very simple. It is a function which
gets the current store:

```typescript
function withSum(a: Signal<number>, b: Signal<number>) {
  return signalStoreFeature(
    withComputed(() => ({
      sum: computed(() => a() + b())
    }))
  );
}

signalStore(
  withState({ a: 1, b: 2 }),
  withFeatureFactory((store) => withSum(store.a, store.b))
);
```