feat: add mutation api
rxMutation creates a mutation function in the store delegating to a
function/method returning an observable and takes care of race conditions
by internally using mergeMap (default), switchMap, concatMap, or
exhaustMap.

mutation also creates an async mutation function delegating to a
function/method returning an observable. When the async function is awaited
without throwing an exception, the caller can safely assume that it has
been executed. However, here, the consumer is responsible for preventing
race conditions.