feat: add `withResource`
This adds the implementation for `withResource` for both named
and single resources.

The implementation is close to the PR in NgRx but doesn't use any
internal types.

Because of a bug in `withLinkedState` https://github.com/ngrx/platform/issues/4931,
we currently have no override protected for `value`.

For more information, check out the docs.

Co-authored-by: Michael Small <michaelsmalldev@gmail.com>