chore: v21 + `moduleResolution: "bundler"`
A few changes had to be done by hand

#1 `moduleResolution` change
https://www.typescriptlang.org/tsconfig/#moduleResolution

The "node" option for the module resolution was so old
that it was called "node" and not even "node10" like
in the TS docs. I looked at a v21 Stackblitz project and
Jason Warner's v21 NX project and they both
specified "bundler" instead.

#2 Narrow imports

Imports of ngrx types had to be made less specific

#3 Stricter linting

Mostly things that looked intentional that had to be ignored

#4 Remove `tsconfig` from test config files v21 nx added

Broke imports, I can see now that they were consciously omitted